Added dependency consideration

+++ utilities/pipelines/resourceRemoval/helper/Remove-VirtualMachine.ps1
@@ -41,6 +41,7 @@ function Remove-VirtualMachine {
 
         # Load helper
         . (Join-Path $PSScriptRoot 'Remove-Resource.ps1')
+        . (Join-Path $PSScriptRoot 'Get-DependencyResourceNames.ps1')
     }
 
     process {
@@ -63,25 +64,48 @@ function Remove-VirtualMachine {
         }
 
         $unorderedResourceIds = $deployments.TargetResource | Where-Object { $_ -and $_ -notmatch '/deployments/' }
-        $vmName = ($unorderedResourceIds | Where-Object { $_ -match '/virtualMachines/' }).Split('/')[-1]
+        $resourcesToRemove = [System.Collections.ArrayList] @()
 
-        # Fetch all resources that match the VM
-        $allResources = Get-AzResource -ResourceGroupName $resourceGroupName -Name "$vmName*"
 
-        # Sort ascending as we need to remove the VM first
-        $orderedResourceIds = $allResources | Sort-Object -Property { $_.ResourceId.Split('/').Count }
-        $resourcesToRemove = $orderedResourceIds | ForEach-Object {
-            @{
-                resourceId = $_.ResourceId
-                name       = $_.Name
-                type       = $_.Type
+        # Handle VM resources
+        $vmIds = $unorderedResourceIds | Where-Object { $_ -match '/virtualMachines/' }
+        if ($vmIds.Count -gt 0) {
+            $vmName = ($unorderedResourceIds | Where-Object { $_ -match '/virtualMachines/' }).Split('/')[-1]
+
+            # Fetch all resources that match the VM
+            $allResources = Get-AzResource -ResourceGroupName $resourceGroupName -Name "$vmName*"
+
+            # Sort ascending as we need to remove the VM first
+            $orderedResourceIds = $allResources | Sort-Object -Property { $_.ResourceId.Split('/').Count }
+            $resourcesToRemove += $orderedResourceIds | ForEach-Object {
+                @{
+                    resourceId = $_.ResourceId
+                    name       = $_.Name
+                    type       = $_.Type
+                }
             }
         }
 
+        # Handle non-vm resources that are deployed with the VM
+        $otherIds = $unorderedResourceIds | Where-Object { $_ -notmatch '/virtualMachines/' }
+        foreach ($otherResourceId in $otherIds) {
+            $resourcesToRemove += @{
+                resourceId = $otherResourceId
+                name       = $otherResourceId.Split('/')[-1]
+                type       = $otherResourceId.Split('/')[6..7] -join '/'
+            }
+        }
+
+        # Filter all dependency resources
+        $dependencyResourceNames = Get-DependencyResourceNames
+        $resourcesToRemove = $resourcesToRemove | Where-Object { $_.Name -notin $dependencyResourceNames }
+
         # Remove resources
         # ----------------
-        if ($PSCmdlet.ShouldProcess(('[{0}] resources' -f $resourcesToRemove.Count), 'Remove')) {
-            Remove-Resource -resourceToRemove $resourcesToRemove -Verbose
+        if ($resourcesToRemove.count -gt 0) {
+            if ($PSCmdlet.ShouldProcess(('[{0}] resources' -f $resourcesToRemove.Count), 'Remove')) {
+                Remove-Resource -resourceToRemove $resourcesToRemove -Verbose
+            }
         }
     }
